This shows the zipped package, for web scraping directly in AWS Lambda
This holds all python and beautufl soup modules needed for basic webscraping. This was created in a pipenv virtual Python 3.8 environment in the command terminal, combined with the lambda_function_2.py, zipped together, and uploaded directly to AWS Lambda. This allowed for the running of webscraping for cheese characteristics. Output log to CloudWatch on AWS.